[cargo-nextest] execute integration tests as subprocesses
Previously, we were executing cargo-nextest in-process due to the fact
that we can't replace binaries on Windows while they're running. To
solve that issue, create a separate crate with a separate
"cargo-nextest-dup" binary that's only used for integration tests, and
is otherwise identical to the cargo-nextest binary.

We're going to need this for upcoming double-spawn work.
