+++ cargo-nextest/Cargo.toml
@@ -40,11 +40,6 @@ serde_json = "1.0.87"
 thiserror = "1.0.37"
 nextest-workspace-hack = { version = "0.1", path = "../workspace-hack" }
 
-[dev-dependencies]
-cfg-if = "1.0.0"
-tempfile = "3.3.0"
-regex = "1.6.0"
-
 [features]
 default = ["default-no-update", "self-update"]
 # Perform self-updates (enabled by default)
