+++ integration-tests/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "integration-tests"
+version = "0.1.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+
+[[bin]]
+name = "cargo-nextest-dup"
+path = "test-helpers/cargo-nextest-dup.rs"
+
+[dependencies]
+cargo-nextest = { path = "../cargo-nextest" }
+# we don't use the tracing support
+color-eyre = { version = "0.6.2", default-features = false }
+clap = { version = "4.0.19", features = ["derive", "env"] }
+enable-ansi-support = "0.1.2"
+nextest-workspace-hack = { version = "0.1", path = "../workspace-hack" }
+
+[dev-dependencies]
+camino = "1.1.1"
+cfg-if = "1.0.0"
+pathdiff = { version = "0.2.1", features = ["camino"] }
+nextest-metadata = { version = "=0.7.0", path = "../nextest-metadata" }
+once_cell = "1.16.0"
+tempfile = "3.3.0"
+regex = "1.6.0"
+serde_json = "1.0.87"
